workflows: don't expect repo to be valid before timestamp
Right now the timestamp is expired already (as tuf API changed
so timestamp actions failed for a while) and timestamp action
now fails to run because of that.

Ideally we'd recognise this situation form other "invalid repo"
situations but currently we are not able to:i Don't validate before
timestamp, just run timestamp and see if the result is valid.